メモ機能をコンポーネント化
